iface: move filters into a dedicated module
